Co-creation-projects/huailishang-AgentPlatformBase/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.0
python-dotenv>=1.0
requests>=2.31
hello-agents==0.2.9
openai>=1.12.0
tavily-python>=0.5.0
ddgs>=9.6.1
loguru>=0.7.3
typing_extensions>=4.8.0
